change typedef VA to uint64_t

+++ mcsema/Arch/Arch.h
@@ -74,7 +74,7 @@ class MCInst;
 
 }  // namespace llvm
 
-typedef uintptr_t VA;
+typedef uint64_t VA;
 
 enum SystemArchType {
   _X86_,
